+++ pkgs/cross-compiling/cross-pkgs.nix
@@ -1,6 +1,6 @@
 # Copyright (c) 2019-2024, see AUTHORS. Licensed under MIT License, see LICENSE.
 
-{ callPackage, nixpkgs, system }:
+{ callPackage, nixpkgs }:
 
 let
   args = callPackage ./cross-pkgs-args.nix { };